msm: camera: Remove wingtech camera dependencies
* May cause mismatches between userspace and kernel.
* Thanks to kumajaya for the heads up.
