+++ drivers/media/platform/msm/Kconfig
@@ -33,13 +33,6 @@ config MSMB_CAMERA_DEBUG
 	---help---
 	  Enable printk() debug for msm camera 2.0
 
-config WINGTECH_CAMERA
-	bool "Wingtech's variant of the camera stack"
-	depends on ARCH_MSM && VIDEO_V4L2 && I2C && MSMB_CAMERA && MACH_XIAOMI
-	default n
-	---help---
-          Enable the old Wingtech camera code
-
 if MSMB_CAMERA
 source "drivers/media/platform/msm/camera_v2/Kconfig"
 endif # MSMB_CAMERA
