+++ include/media/camera2.h
@@ -14,11 +14,7 @@
 #ifndef __CAMERA2_H__
 #define __CAMERA2_H__
 
-#ifdef CONFIG_WINGTECH_CAMERA
-#include <media/msm_cam_sensor_wingtech.h>
-#else
 #include <media/msm_cam_sensor.h>
-#endif
 #include <linux/interrupt.h>
 #include <linux/of_platform.h>
 #include <linux/of_device.h>
@@ -96,9 +92,6 @@ struct msm_camera_sensor_board_info {
 	const char *sensor_name;
 	const char *eeprom_name;
 	const char *actuator_name;
-#ifndef CONFIG_WINGTECH_CAMERA
-	const char *ois_name;
-#endif
 	struct msm_camera_slave_info *slave_info;
 	struct msm_camera_csi_lane_params *csi_lane_params;
 	struct msm_camera_sensor_strobe_flash_data *strobe_flash_data;
