+++ include/media/msm_cam_sensor.h
@@ -1,34 +1,46 @@
-#ifdef CONFIG_WINGTECH_CAMERA
-#include <media/msm_cam_sensor_wingtech.h>
-#else
-
 #ifndef __LINUX_MSM_CAM_SENSOR_H
 #define __LINUX_MSM_CAM_SENSOR_H
 
 #ifdef MSM_CAMERA_BIONIC
 #include <sys/types.h>
 #endif
-
-#include <media/msm_camsensor_sdk.h>
-
 #include <linux/types.h>
+#include <linux/v4l2-mediabus.h>
 #include <linux/i2c.h>
-#ifdef CONFIG_COMPAT
-#include <linux/compat.h>
-#endif
 
 #define I2C_SEQ_REG_SETTING_MAX   5
+#define I2C_SEQ_REG_DATA_MAX      20
+#define MAX_CID                   16
 
 #define MSM_SENSOR_MCLK_8HZ   8000000
 #define MSM_SENSOR_MCLK_16HZ  16000000
 #define MSM_SENSOR_MCLK_24HZ  24000000
 
-#define MAX_SENSOR_NAME 32
-#define MAX_ACTUATOR_AF_TOTAL_STEPS 1024
+#define GPIO_OUT_LOW          (0 << 1)
+#define GPIO_OUT_HIGH         (1 << 1)
 
-#define MAX_OIS_MOD_NAME_SIZE 32
-#define MAX_OIS_NAME_SIZE 32
-#define MAX_OIS_REG_SETTINGS 800
+#define CSI_EMBED_DATA        0x12
+#define CSI_RESERVED_DATA_0   0x13
+#define CSI_YUV422_8          0x1E
+#define CSI_RAW8              0x2A
+#define CSI_RAW10             0x2B
+#define CSI_RAW12             0x2C
+
+#define CSI_DECODE_6BIT         0
+#define CSI_DECODE_8BIT         1
+#define CSI_DECODE_10BIT        2
+#define CSI_DECODE_DPCM_10_8_10 5
+
+#define MAX_SENSOR_NAME 32
+
+#define MAX_ACT_MOD_NAME_SIZE 32
+#define MAX_ACT_NAME_SIZE 32
+#define NUM_ACTUATOR_DIR 2
+#define MAX_ACTUATOR_SCENARIO 8
+#define MAX_ACTUATOR_REGION 5
+#define MAX_ACTUATOR_INIT_SET 12
+#define MAX_ACTUATOR_REG_TBL_SIZE 8
+#define MAX_ACTUATOR_AF_TOTAL_STEPS 1024
 
 #define MOVE_NEAR 0
 #define MOVE_FAR  1
@@ -43,10 +55,68 @@
 
 #define MAX_LED_TRIGGERS 3
 
+enum sensor_stats_type {
+	YRGB,
+	YYYY,
+};
+
 enum flash_type {
 	LED_FLASH = 1,
 	STROBE_FLASH,
 	GPIO_FLASH
+};
+
+enum msm_camera_i2c_reg_addr_type {
+	MSM_CAMERA_I2C_BYTE_ADDR = 1,
+	MSM_CAMERA_I2C_WORD_ADDR,
+	MSM_CAMERA_I2C_3B_ADDR,
+	MSM_CAMERA_I2C_ADDR_TYPE_MAX,
+};
+
+enum msm_camera_i2c_data_type {
+	MSM_CAMERA_I2C_BYTE_DATA = 1,
+	MSM_CAMERA_I2C_WORD_DATA,
+	MSM_CAMERA_I2C_SET_BYTE_MASK,
+	MSM_CAMERA_I2C_UNSET_BYTE_MASK,
+	MSM_CAMERA_I2C_SET_WORD_MASK,
+	MSM_CAMERA_I2C_UNSET_WORD_MASK,
+	MSM_CAMERA_I2C_SET_BYTE_WRITE_MASK_DATA,
+	MSM_CAMERA_I2C_DATA_TYPE_MAX,
+};
+
+enum msm_sensor_power_seq_type_t {
+	SENSOR_CLK,
+	SENSOR_GPIO,
+	SENSOR_VREG,
+	SENSOR_I2C_MUX,
+};
+
+enum msm_sensor_clk_type_t {
+	SENSOR_CAM_MCLK,
+	SENSOR_CAM_CLK,
+	SENSOR_CAM_CLK_MAX,
+};
+
+enum msm_sensor_power_seq_gpio_t {
+	SENSOR_GPIO_RESET,
+	SENSOR_GPIO_STANDBY,
+	SENSOR_GPIO_AF_PWDM,
+	SENSOR_GPIO_VIO,
+	SENSOR_GPIO_VANA,
+	SENSOR_GPIO_VDIG,
+	SENSOR_GPIO_VAF,
+	SENSOR_GPIO_FL_EN,
+	SENSOR_GPIO_FL_NOW,
+	SENSOR_GPIO_FL_RESET,
+	SENSOR_GPIO_MAX,
+};
+
+enum msm_camera_vreg_name_t {
+	CAM_VDIG,
+	CAM_VIO,
+	CAM_VANA,
+	CAM_VAF,
+	CAM_VREG_MAX,
 };
 
 enum msm_sensor_resolution_t {
@@ -79,7 +149,6 @@ enum sensor_sub_module_t {
 	SUB_MODULE_CSID_3D,
 	SUB_MODULE_CSIPHY,
 	SUB_MODULE_CSIPHY_3D,
-	SUB_MODULE_OIS,
 	SUB_MODULE_MAX,
 };
 
@@ -152,8 +221,9 @@ enum csiphy_cfg_type_t {
 };
 
 enum camera_vreg_type {
-	VREG_TYPE_DEFAULT,
-	VREG_TYPE_CUSTOM,
+	REG_LDO,
+	REG_VS,
+	REG_GPIO,
 };
 
 enum sensor_af_t {
@@ -161,10 +231,72 @@ enum sensor_af_t {
 	SENSOR_AF_NOT_FOCUSSED,
 };
 
+struct msm_sensor_power_setting {
+	enum msm_sensor_power_seq_type_t seq_type;
+	uint16_t seq_val;
+	long config_val;
+	uint16_t delay;
+	void *data[10];
+};
+
+struct msm_sensor_power_setting_array {
+	struct msm_sensor_power_setting *power_setting;
+	uint16_t size;
+	struct msm_sensor_power_setting *power_down_setting;
+	uint16_t size_down;
+};
+
+struct msm_sensor_id_info_t {
+	uint16_t sensor_id_reg_addr;
+	uint16_t sensor_id;
+};
+
+enum msm_sensor_camera_id_t {
+	CAMERA_0,
+	CAMERA_1,
+	CAMERA_2,
+	CAMERA_3,
+	MAX_CAMERAS,
+};
+
 enum cci_i2c_master_t {
 	MASTER_0,
 	MASTER_1,
 	MASTER_MAX,
+};
+
+enum i2c_freq_mode_t {
+	I2C_STANDARD_MODE,
+	I2C_FAST_MODE,
+	I2C_CUSTOM_MODE,
+	I2C_MAX_MODES,
+};
+
+struct msm_camera_i2c_reg_array {
+	uint16_t reg_addr;
+	uint16_t reg_data;
+	uint32_t delay;
+};
+
+struct msm_camera_i2c_reg_setting {
+	struct msm_camera_i2c_reg_array *reg_setting;
+	uint16_t size;
+	enum msm_camera_i2c_reg_addr_type addr_type;
+	enum msm_camera_i2c_data_type data_type;
+	uint16_t delay;
+};
+
+struct msm_camera_i2c_seq_reg_array {
+	uint16_t reg_addr;
+	uint8_t reg_data[I2C_SEQ_REG_DATA_MAX];
+	uint16_t reg_data_size;
+};
+
+struct msm_camera_i2c_seq_reg_setting {
+	struct msm_camera_i2c_seq_reg_array *reg_setting;
+	uint16_t size;
+	enum msm_camera_i2c_reg_addr_type addr_type;
+	uint16_t delay;
 };
 
 struct msm_camera_i2c_array_write_config {
@@ -176,13 +308,38 @@ struct msm_camera_i2c_read_config {
 	uint16_t slave_addr;
 	uint16_t reg_addr;
 	enum msm_camera_i2c_data_type data_type;
-	uint16_t data;
+	uint16_t *data;
+};
+
+struct msm_camera_csid_vc_cfg {
+	uint8_t cid;
+	uint8_t dt;
+	uint8_t decode_format;
+};
+
+struct msm_camera_csid_lut_params {
+	uint8_t num_cid;
+	struct msm_camera_csid_vc_cfg *vc_cfg[MAX_CID];
+};
+
+struct msm_camera_csid_params {
+	uint8_t lane_cnt;
+	uint16_t lane_assign;
+	uint8_t phy_sel;
+	struct msm_camera_csid_lut_params lut_params;
+};
+
+struct msm_camera_csiphy_params {
+	uint8_t lane_cnt;
+	uint8_t settle_cnt;
+	uint16_t lane_mask;
+	uint8_t combo_mode;
+	uint8_t csid_core;
 };
 
 struct msm_camera_csi2_params {
 	struct msm_camera_csid_params csid_params;
 	struct msm_camera_csiphy_params csiphy_params;
-	uint8_t csi_clk_scale_enable;
 };
 
 struct msm_camera_csi_lane_params {
@@ -194,15 +351,20 @@ struct csi_lane_params_t {
 	uint16_t csi_lane_assign;
 	uint8_t csi_lane_mask;
 	uint8_t csi_if;
-	int8_t csid_core[2];
+	uint8_t csid_core[2];
 	uint8_t csi_phy_sel;
+};
+
+enum camb_position_t {
+	BACK_CAMERA_B,
+	FRONT_CAMERA_B,
+	INVALID_CAMERA_B,
 };
 
 struct msm_sensor_info_t {
 	char     sensor_name[MAX_SENSOR_NAME];
 	uint32_t session_id;
 	int32_t  subdev_id[SUB_MODULE_MAX];
-	int32_t  subdev_intf[SUB_MODULE_MAX];
 	uint8_t  is_mount_angle_valid;
 	uint32_t sensor_mount_angle;
 	int modes_supported;
@@ -211,12 +373,40 @@ struct msm_sensor_info_t {
 
 struct camera_vreg_t {
 	const char *reg_name;
+	enum camera_vreg_type type;
 	int min_voltage;
 	int max_voltage;
 	int op_mode;
 	uint32_t delay;
-	const char *custom_vreg_name;
-	enum camera_vreg_type type;
+};
+
+enum camerab_mode_t {
+	CAMERA_MODE_2D_B = (1<<0),
+	CAMERA_MODE_3D_B = (1<<1),
+	CAMERA_MODE_INVALID = (1<<2),
+};
+
+struct msm_sensor_init_params {
+	/* mask of modes supported: 2D, 3D */
+	int                 modes_supported;
+	/* sensor position: front, back */
+	enum camb_position_t position;
+	/* sensor mount angle */
+	uint32_t            sensor_mount_angle;
+};
+
+struct msm_camera_sensor_slave_info {
+	char sensor_name[32];
+	char eeprom_name[32];
+	char actuator_name[32];
+	enum msm_sensor_camera_id_t camera_id;
+	uint16_t slave_addr;
+	enum i2c_freq_mode_t i2c_freq_mode;
+	enum msm_camera_i2c_reg_addr_type addr_type;
+	struct msm_sensor_id_info_t sensor_id_info;
+	struct msm_sensor_power_setting_array power_setting_array;
+	uint8_t  is_init_params_valid;
+	struct msm_sensor_init_params sensor_init_params;
 };
 
 struct sensorb_cfg_data {
@@ -284,100 +474,6 @@ struct msm_eeprom_cfg_data {
 	} cfg;
 };
 
-#ifdef CONFIG_COMPAT
-struct msm_sensor_power_setting32 {
-	enum msm_sensor_power_seq_type_t seq_type;
-	uint16_t seq_val;
-	compat_uint_t config_val;
-	uint16_t delay;
-	compat_uptr_t data[10];
-};
-
-struct msm_sensor_power_setting_array32 {
-	struct msm_sensor_power_setting32 power_setting_a[MAX_POWER_CONFIG];
-	compat_uptr_t power_setting;
-	uint16_t size;
-	struct msm_sensor_power_setting32
-		power_down_setting_a[MAX_POWER_CONFIG];
-	compat_uptr_t power_down_setting;
-	uint16_t size_down;
-};
-
-struct msm_camera_sensor_slave_info32 {
-	char sensor_name[32];
-	char eeprom_name[32];
-	char actuator_name[32];
-	char ois_name[32];
-	char flash_name[32];
-	enum msm_sensor_camera_id_t camera_id;
-	uint16_t slave_addr;
-	enum i2c_freq_mode_t i2c_freq_mode;
-	enum msm_camera_i2c_reg_addr_type addr_type;
-	struct msm_sensor_id_info_t sensor_id_info;
-	struct msm_sensor_power_setting_array32 power_setting_array;
-	uint8_t  is_init_params_valid;
-	struct msm_sensor_init_params sensor_init_params;
-	uint8_t is_flash_supported;
-	enum msm_sensor_output_format_t output_format;
-};
-
-struct msm_camera_csid_lut_params32 {
-	uint8_t num_cid;
-	struct msm_camera_csid_vc_cfg vc_cfg_a[MAX_CID];
-	compat_uptr_t vc_cfg[MAX_CID];
-};
-
-struct msm_camera_csid_params32 {
-	uint8_t lane_cnt;
-	uint16_t lane_assign;
-	uint8_t phy_sel;
-	uint32_t csi_clk;
-	struct msm_camera_csid_lut_params32 lut_params;
-};
-
-struct msm_camera_csi2_params32 {
-	struct msm_camera_csid_params32 csid_params;
-	struct msm_camera_csiphy_params csiphy_params;
-	uint8_t csi_clk_scale_enable;
-};
-
-struct csid_cfg_data32 {
-	enum csid_cfg_type_t cfgtype;
-	union {
-		uint32_t csid_version;
-		compat_uptr_t csid_params;
-	} cfg;
-};
-
-struct eeprom_read_t32 {
-	compat_uptr_t dbuffer;
-	uint32_t num_bytes;
-};
-
-struct eeprom_write_t32 {
-	compat_uptr_t dbuffer;
-	uint32_t num_bytes;
-};
-
-struct msm_eeprom_cfg_data32 {
-	enum eeprom_cfg_type_t cfgtype;
-	uint8_t is_supported;
-	union {
-		char eeprom_name[MAX_SENSOR_NAME];
-		struct eeprom_get_t get_data;
-		struct eeprom_read_t32 read_data;
-		struct eeprom_write_t32 write_data;
-	} cfg;
-};
-
-struct msm_camera_i2c_seq_reg_setting32 {
-	compat_uptr_t reg_setting;
-	uint16_t size;
-	enum msm_camera_i2c_reg_addr_type addr_type;
-	uint16_t delay;
-};
-#endif
-
 enum msm_sensor_cfg_type_t {
 	CFG_SET_SLAVE_INFO,
 	CFG_SLAVE_READ_I2C,
@@ -418,39 +514,47 @@ enum msm_actuator_cfg_type_t {
 	CFG_ACTUATOR_INIT,
 };
 
-enum msm_ois_cfg_type_t {
-	CFG_OIS_INIT,
-	CFG_OIS_POWERDOWN,
-	CFG_OIS_POWERUP,
-	CFG_OIS_CONTROL,
-	CFG_OIS_I2C_WRITE_SEQ_TABLE,
+enum actuator_type {
+	ACTUATOR_VCM,
+	ACTUATOR_PIEZO,
 };
 
-enum msm_ois_i2c_operation {
-	MSM_OIS_WRITE = 0,
-	MSM_OIS_POLL,
+enum msm_actuator_data_type {
+	MSM_ACTUATOR_BYTE_DATA = 1,
+	MSM_ACTUATOR_WORD_DATA,
 };
 
-struct reg_settings_ois_t {
+enum msm_actuator_addr_type {
+	MSM_ACTUATOR_BYTE_ADDR = 1,
+	MSM_ACTUATOR_WORD_ADDR,
+};
+
+enum msm_actuator_i2c_operation {
+	MSM_ACT_WRITE = 0,
+	MSM_ACT_POLL,
+};
+
+struct reg_settings_t {
 	uint16_t reg_addr;
-	enum msm_camera_i2c_reg_addr_type addr_type;
-	uint32_t reg_data;
-	enum msm_camera_i2c_data_type data_type;
-	enum msm_ois_i2c_operation i2c_operation;
+	enum msm_actuator_addr_type addr_type;
+	uint16_t reg_data;
+	enum msm_actuator_data_type data_type;
+	enum msm_actuator_i2c_operation i2c_operation;
 	uint32_t delay;
 };
 
-struct msm_ois_params_t {
-	uint16_t data_size;
-	uint16_t setting_size;
-	uint32_t i2c_addr;
-	enum msm_camera_i2c_reg_addr_type i2c_addr_type;
-	enum msm_camera_i2c_data_type i2c_data_type;
-	struct reg_settings_ois_t *settings;
+struct region_params_t {
+	/* [0] = ForwardDirection Macro boundary
+	   [1] = ReverseDirection Inf boundary
+	*/
+	uint16_t step_bound[2];
+	uint16_t code_per_step;
 };
 
-struct msm_ois_set_info_t {
-	struct msm_ois_params_t ois_params;
+struct damping_params_t {
+	uint32_t damping_step;
+	uint32_t damping_delay;
+	uint32_t hw_params;
 };
 
 struct msm_actuator_move_params_t {
@@ -470,13 +574,6 @@ struct msm_actuator_tuning_params_t {
 	struct region_params_t *region_params;
 };
 
-struct park_lens_data_t {
-	uint32_t damping_step;
-	uint32_t damping_delay;
-	uint32_t hw_params;
-	uint32_t max_step;
-};
-
 struct msm_actuator_params_t {
 	enum actuator_type act_type;
 	uint8_t reg_tbl_size;
@@ -487,7 +584,6 @@ struct msm_actuator_params_t {
 	enum msm_actuator_data_type i2c_data_type;
 	struct msm_actuator_reg_params_t *reg_tbl_params;
 	struct reg_settings_t *init_settings;
-	struct park_lens_data_t park_lens;
 };
 
 struct msm_actuator_set_info_t {
@@ -522,13 +618,6 @@ enum af_camera_name {
 	ACTUATOR_WEB_CAM_2,
 };
 
-struct msm_ois_cfg_data {
-	int cfgtype;
-	union {
-		struct msm_ois_set_info_t set_info;
-		struct msm_camera_i2c_seq_reg_setting *settings;
-	} cfg;
-};
 
 struct msm_actuator_set_position_t {
 	uint16_t number_of_steps;
@@ -548,6 +637,19 @@ struct msm_actuator_cfg_data {
 	} cfg;
 };
 
+enum msm_actuator_write_type {
+	MSM_ACTUATOR_WRITE_HW_DAMP,
+	MSM_ACTUATOR_WRITE_DAC,
+};
+
+struct msm_actuator_reg_params_t {
+	enum msm_actuator_write_type reg_write_type;
+	uint32_t hw_mask;
+	uint16_t reg_addr;
+	uint16_t hw_shift;
+	uint16_t data_shift;
+};
+
 enum msm_camera_led_config_t {
 	MSM_CAMERA_LED_OFF,
 	MSM_CAMERA_LED_LOW,
@@ -558,27 +660,8 @@ enum msm_camera_led_config_t {
 
 struct msm_camera_led_cfg_t {
 	enum msm_camera_led_config_t cfgtype;
-	uint32_t torch_current[MAX_LED_TRIGGERS];
-	uint32_t flash_current[MAX_LED_TRIGGERS];
-	uint32_t flash_duration[MAX_LED_TRIGGERS];
-};
-
-struct msm_flash_init_info_t {
-	enum msm_flash_driver_type flash_driver_type;
-	struct msm_sensor_power_setting_array *power_setting_array;
-	struct msm_camera_i2c_reg_setting_array *settings;
-};
-
-struct msm_flash_cfg_data_t {
-	enum msm_flash_cfg_type_t cfg_type;
 	uint32_t torch_current;
 	uint32_t flash_current[MAX_LED_TRIGGERS];
-	uint32_t flash_duration[MAX_LED_TRIGGERS];
-	union {
-		struct msm_flash_init_info_t *flash_init_info;
-		struct msm_camera_i2c_reg_setting_array *settings;
-		uint32_t flash_current[MAX_LED_TRIGGERS];
-	} cfg;
 };
 
 /* sensor init structures and enums */
@@ -590,8 +673,6 @@ enum msm_sensor_init_cfg_type_t {
 
 struct sensor_init_cfg_data {
 	enum msm_sensor_init_cfg_type_t cfgtype;
-	struct msm_sensor_info_t        probed_info;
-	char                            entity_name[MAX_SENSOR_NAME];
 	union {
 		void *setting;
 	} cfg;
@@ -627,159 +708,6 @@ struct sensor_init_cfg_data {
 #define VIDIOC_MSM_SENSOR_INIT_CFG \
 	_IOWR('V', BASE_VIDIOC_PRIVATE + 10, struct sensor_init_cfg_data)
 
-#define VIDIOC_MSM_OIS_CFG \
-	_IOWR('V', BASE_VIDIOC_PRIVATE + 11, struct msm_ois_cfg_data)
-
-#define VIDIOC_MSM_FLASH_CFG \
-	_IOWR('V', BASE_VIDIOC_PRIVATE + 13, struct msm_flash_cfg_data_t)
-
-#ifdef CONFIG_COMPAT
-struct msm_camera_i2c_reg_setting32 {
-	compat_uptr_t reg_setting;
-	uint16_t size;
-	enum msm_camera_i2c_reg_addr_type addr_type;
-	enum msm_camera_i2c_data_type data_type;
-	uint16_t delay;
-	enum msm_camera_qup_i2c_write_batch_t qup_i2c_batch;
-};
-
-struct msm_actuator_tuning_params_t32 {
-	int16_t initial_code;
-	uint16_t pwd_step;
-	uint16_t region_size;
-	uint32_t total_steps;
-	compat_uptr_t region_params;
-};
-
-struct msm_actuator_params_t32 {
-	enum actuator_type act_type;
-	uint8_t reg_tbl_size;
-	uint16_t data_size;
-	uint16_t init_setting_size;
-	uint32_t i2c_addr;
-	enum msm_actuator_addr_type i2c_addr_type;
-	enum msm_actuator_data_type i2c_data_type;
-	compat_uptr_t reg_tbl_params;
-	compat_uptr_t init_settings;
-	struct park_lens_data_t park_lens;
-};
-
-struct msm_actuator_set_info_t32 {
-	struct msm_actuator_params_t32 actuator_params;
-	struct msm_actuator_tuning_params_t32 af_tuning_params;
-};
-
-struct sensor_init_cfg_data32 {
-	enum msm_sensor_init_cfg_type_t cfgtype;
-	struct msm_sensor_info_t        probed_info;
-	char                            entity_name[MAX_SENSOR_NAME];
-	union {
-		compat_uptr_t setting;
-	} cfg;
-};
-
-struct msm_actuator_move_params_t32 {
-	int8_t dir;
-	int8_t sign_dir;
-	int16_t dest_step_pos;
-	int32_t num_steps;
-	uint16_t curr_lens_pos;
-	compat_uptr_t ringing_params;
-};
-
-struct msm_actuator_cfg_data32 {
-	int cfgtype;
-	uint8_t is_af_supported;
-	union {
-		struct msm_actuator_move_params_t32 move;
-		struct msm_actuator_set_info_t32 set_info;
-		struct msm_actuator_get_info_t get_info;
-		struct msm_actuator_set_position_t setpos;
-		enum af_camera_name cam_name;
-	} cfg;
-};
-
-struct csiphy_cfg_data32 {
-	enum csiphy_cfg_type_t cfgtype;
-	union {
-		compat_uptr_t csiphy_params;
-		compat_uptr_t csi_lane_params;
-	} cfg;
-};
-
-struct sensorb_cfg_data32 {
-	int cfgtype;
-	union {
-		struct msm_sensor_info_t      sensor_info;
-		struct msm_sensor_init_params sensor_init_params;
-		compat_uptr_t                 setting;
-	} cfg;
-};
-
-struct msm_ois_params_t32 {
-	uint16_t data_size;
-	uint16_t setting_size;
-	uint32_t i2c_addr;
-	enum msm_camera_i2c_reg_addr_type i2c_addr_type;
-	enum msm_camera_i2c_data_type i2c_data_type;
-	compat_uptr_t settings;
-};
-
-struct msm_ois_set_info_t32 {
-	struct msm_ois_params_t32 ois_params;
-};
-
-struct msm_ois_cfg_data32 {
-	int cfgtype;
-	union {
-		struct msm_ois_set_info_t32 set_info;
-		compat_uptr_t settings;
-	} cfg;
-};
-
-struct msm_flash_init_info_t32 {
-	enum msm_flash_driver_type flash_driver_type;
-	compat_uptr_t power_setting_array;
-	compat_uptr_t settings;
-};
-
-struct msm_flash_cfg_data_t32 {
-	enum msm_flash_cfg_type_t cfg_type;
-	uint32_t torch_current;
-	uint32_t flash_current[MAX_LED_TRIGGERS];
-	uint32_t flash_duration[MAX_LED_TRIGGERS];
-	union {
-		compat_uptr_t flash_init_info;
-		compat_uptr_t settings;
-		uint32_t flash_current[MAX_LED_TRIGGERS];
-	} cfg;
-};
-
-#define VIDIOC_MSM_ACTUATOR_CFG32 \
-	_IOWR('V', BASE_VIDIOC_PRIVATE + 6, struct msm_actuator_cfg_data32)
-
-#define VIDIOC_MSM_SENSOR_INIT_CFG32 \
-	_IOWR('V', BASE_VIDIOC_PRIVATE + 10, struct sensor_init_cfg_data32)
-
-#define VIDIOC_MSM_CSIPHY_IO_CFG32 \
-	_IOWR('V', BASE_VIDIOC_PRIVATE + 4, struct csiphy_cfg_data32)
-
-#define VIDIOC_MSM_SENSOR_CFG32 \
-	_IOWR('V', BASE_VIDIOC_PRIVATE + 1, struct sensorb_cfg_data32)
-
-#define VIDIOC_MSM_EEPROM_CFG32 \
-	_IOWR('V', BASE_VIDIOC_PRIVATE + 8, struct msm_eeprom_cfg_data32)
-
-#define VIDIOC_MSM_OIS_CFG32 \
-	_IOWR('V', BASE_VIDIOC_PRIVATE + 11, struct msm_ois_cfg_data32)
-
-#define VIDIOC_MSM_CSID_IO_CFG32 \
-	_IOWR('V', BASE_VIDIOC_PRIVATE + 5, struct csid_cfg_data32)
-
-#define VIDIOC_MSM_FLASH_CFG32 \
-	_IOWR('V', BASE_VIDIOC_PRIVATE + 13, struct msm_flash_cfg_data_t32)
-#endif
+#define MSM_V4L2_PIX_FMT_META v4l2_fourcc('M', 'E', 'T', 'A') /* META */
 
 #endif /* __LINUX_MSM_CAM_SENSOR_H */
-
-#endif /* CONFIG_WINGTECH_CAMERA */
