+++ include/media/msmb_ispif.h
@@ -101,9 +101,6 @@ enum ispif_cfg_type_t {
 	ISPIF_INIT,
 	ISPIF_CFG,
 	ISPIF_START_FRAME_BOUNDARY,
-#ifndef CONFIG_WINGTECH_CAMERA
-	ISPIF_RESTART_FRAME_BOUNDARY,
-#endif
 	ISPIF_STOP_FRAME_BOUNDARY,
 	ISPIF_STOP_IMMEDIATELY,
 	ISPIF_RELEASE,
