+++ include/media/msmb_pproc.h
@@ -1,15 +1,8 @@
-#ifdef CONFIG_WINGTECH_CAMERA
-#include <media/msmb_pproc_wingtech.h>
-#else
-
 #ifndef __MSMB_PPROC_H
 #define __MSMB_PPROC_H
 
 #ifdef MSM_CAMERA_BIONIC
 #include <sys/types.h>
-#endif
-#ifdef CONFIG_COMPAT
-#include <linux/compat.h>
 #endif
 #include <linux/videodev2.h>
 #include <linux/types.h>
@@ -20,10 +13,10 @@
 
 #define MAX_NUM_CPP_STRIPS 8
 #define MSM_CPP_MAX_NUM_PLANES 3
-#define MSM_CPP_MIN_FRAME_LENGTH 13
-#define MSM_CPP_MAX_FRAME_LENGTH 2048
+#define MSM_CPP_MAX_FRAME_LENGTH 1024
 #define MSM_CPP_MAX_FW_NAME_LEN 32
 #define MAX_FREQ_TBL 10
+
 
 enum msm_cpp_frame_type {
 	MSM_CPP_OFFLINE_FRAME,
@@ -107,18 +100,18 @@ struct msm_cpp_frame_strip_info {
 	uint32_t output_bytes_per_pixel;
 	uint32_t temporal_bytes_per_pixel;
 
-	uint32_t source_address[2];
-	uint32_t destination_address[2];
+	unsigned int source_address[2];
+	unsigned int destination_address[2];
 	/* source_address[1] is used for CbCR planar
 	 * to CbCr interleaved conversion
 	 */
-	uint32_t temporal_source_address[2];
+	unsigned int temporal_source_address[2];
 	/* destination_address[1] is used for CbCr interleved
 	 * to CbCr planar conversion
 	 */
-	uint32_t temporal_destination_address[2];
-	uint32_t src_stride;
-	uint32_t dst_stride;
+	unsigned int temporal_destination_address[2];
+	unsigned int src_stride;
+	unsigned int dst_stride;
 	uint32_t rotate_270;
 	uint32_t horizontal_flip;
 	uint32_t vertical_flip;
@@ -132,7 +125,7 @@ struct msm_cpp_frame_strip_info {
 };
 
 struct msm_cpp_buffer_info_t {
-	int32_t fd;
+	int fd;
 	uint32_t index;
 	uint32_t offset;
 	uint8_t native_buff;
@@ -154,24 +147,20 @@ struct msm_cpp_frame_info_t {
 	uint32_t client_id;
 	enum msm_cpp_frame_type frame_type;
 	uint32_t num_strips;
-	struct msm_cpp_frame_strip_info __user *strip_info;
+	struct msm_cpp_frame_strip_info *strip_info;
 	uint32_t msg_len;
 	uint32_t *cpp_cmd_msg;
 	int src_fd;
 	int dst_fd;
+	struct ion_handle *src_ion_handle;
+	struct ion_handle *dest_ion_handle;
 	struct timeval in_time, out_time;
-	void __user *cookie;
+	void *cookie;
 	int32_t *status;
 	int32_t duplicate_output;
 	uint32_t duplicate_identity;
 	struct msm_cpp_buffer_info_t input_buffer_info;
 	struct msm_cpp_buffer_info_t output_buffer_info[2];
-	struct msm_cpp_buffer_info_t tnr_scratch_buffer_info[2];
-};
-
-struct msm_cpp_pop_stream_info_t {
-	int32_t frame_id;
-	uint32_t identity;
 };
 
 struct cpp_hw_info {
@@ -195,7 +184,7 @@ struct msm_vpe_frame_strip_info {
 };
 
 struct msm_vpe_buffer_info_t {
-	int32_t fd;
+	int fd;
 	uint32_t index;
 	uint32_t offset;
 	uint8_t native_buff;
@@ -216,8 +205,8 @@ struct msm_vpe_frame_info_t {
 	uint32_t client_id;
 	enum msm_vpe_frame_type frame_type;
 	struct msm_vpe_frame_strip_info strip_info;
-	unsigned long src_fd;
-	unsigned long dst_fd;
+	int src_fd;
+	int dst_fd;
 	struct ion_handle *src_ion_handle;
 	struct ion_handle *dest_ion_handle;
 	unsigned long src_phyaddr;
@@ -234,12 +223,6 @@ struct msm_vpe_frame_info_t {
 struct msm_pproc_queue_buf_info {
 	struct msm_buf_mngr_info buff_mgr_info;
 	uint8_t is_buf_dirty;
-};
-
-struct msm_cpp_clock_settings_t {
-	unsigned long clock_rate;
-	uint64_t avg;
-	uint64_t inst;
 };
 
 #define VIDIOC_MSM_CPP_CFG \
@@ -265,6 +248,7 @@ struct msm_cpp_clock_settings_t {
 
 #define VIDIOC_MSM_CPP_DEQUEUE_STREAM_BUFF_INFO \
 	_IOWR('V', BASE_VIDIOC_PRIVATE + 7, struct msm_camera_v4l2_ioctl_t)
+
 
 #define VIDIOC_MSM_VPE_CFG \
 	_IOWR('V', BASE_VIDIOC_PRIVATE + 8, struct msm_camera_v4l2_ioctl_t)
@@ -302,13 +286,12 @@ struct msm_cpp_clock_settings_t {
 #define VIDIOC_MSM_CPP_IOMMU_DETACH \
 	_IOWR('V', BASE_VIDIOC_PRIVATE + 19, struct msm_camera_v4l2_ioctl_t)
 
-
 #define V4L2_EVENT_CPP_FRAME_DONE  (V4L2_EVENT_PRIVATE_START + 0)
 #define V4L2_EVENT_VPE_FRAME_DONE  (V4L2_EVENT_PRIVATE_START + 1)
 
 struct msm_camera_v4l2_ioctl_t {
 	uint32_t id;
-	size_t len;
+	uint32_t len;
 	int32_t trans_code;
 	void __user *ioctl_ptr;
 };
@@ -337,12 +320,6 @@ struct msm_cpp_frame_info32_t {
 	struct msm_cpp_buffer_info_t tnr_scratch_buffer_info[2];
 };
 
-struct msm_cpp_clock_settings32_t {
-	compat_long_t clock_rate;
-	uint64_t avg;
-	uint64_t inst;
-};
-
 struct msm_cpp_stream_buff_info32_t {
 	uint32_t identity;
 	uint32_t num_buffs;
@@ -353,14 +330,6 @@ struct msm_pproc_queue_buf_info32_t {
 	struct msm_buf_mngr_info32_t buff_mgr_info;
 	uint8_t is_buf_dirty;
 };
-
-struct cpp_hw_info_32_t {
-	uint32_t cpp_hw_version;
-	uint32_t cpp_hw_caps;
-	compat_long_t freq_tbl[MAX_FREQ_TBL];
-	uint32_t freq_tbl_count;
-};
-
 
 #define VIDIOC_MSM_CPP_CFG32 \
 	_IOWR('V', BASE_VIDIOC_PRIVATE, struct msm_camera_v4l2_ioctl32_t)
@@ -407,9 +376,6 @@ struct cpp_hw_info_32_t {
 #define VIDIOC_MSM_CPP_QUEUE_BUF32 \
 	_IOWR('V', BASE_VIDIOC_PRIVATE + 14, struct msm_camera_v4l2_ioctl32_t)
 
-#define VIDIOC_MSM_CPP_APPEND_STREAM_BUFF_INFO32 \
-	_IOWR('V', BASE_VIDIOC_PRIVATE + 15, struct msm_camera_v4l2_ioctl32_t)
-
 #define VIDIOC_MSM_CPP_SET_CLOCK32 \
 	_IOWR('V', BASE_VIDIOC_PRIVATE + 16, struct msm_camera_v4l2_ioctl32_t)
 
@@ -431,5 +397,3 @@ struct msm_camera_v4l2_ioctl32_t {
 #endif
 
 #endif /* __MSMB_PPROC_H */
-
-#endif  /* CONFIG_WINGTECH_CAMERA */
