+++ include/media/videobuf2-core.h
@@ -18,11 +18,7 @@
 #include <linux/videodev2.h>
 #include <linux/dma-buf.h>
 
-#ifdef CONFIG_WINGTECH_CAMERA
 #define VB2_MAX_FRAME  32
-#else
-#define VB2_MAX_FRAME  64
-#endif
 
 struct vb2_alloc_ctx;
 struct vb2_fileio_data;
