+++ include/uapi/media/Kbuild
@@ -14,10 +14,8 @@ header-y += msmb_camera.h
 header-y += msm_cam_sensor.h
 header-y += msm_camsensor_sdk.h
 header-y += msmb_isp.h
-header-y += msmb_isp_wingtech.h
 header-y += msmb_ispif.h
 header-y += msmb_generic_buf_mgr.h
 header-y += msmb_pproc.h
-header-y += msmb_pproc_wingtech.h
 header-y += msm_vpu.h
 header-y += msm_fd.h
