+++ include/uapi/media/msm_camera.h
@@ -1609,9 +1609,7 @@ struct msm_camera_csi_lane_params {
 
 struct camera_vreg_t {
 	const char *reg_name;
-#ifdef CONFIG_WINGTECH_CAMERA
 	enum camera_vreg_type type;
-#endif
 	int min_voltage;
 	int max_voltage;
 	int op_mode;
